requirements.txt
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
idna==3.10
pycparser==2.22
PyJWT==1.7.1
pyjwt-rsa==0.4.0
requests==2.32.3
urllib3==2.4.0
